Move interfaces to categorized index files in subfolders auth, dto, and ui.
