+++ src/utils/interfaces/index.ts
@@ -1,3 +1,3 @@
 export * from "./dto";
-export * from "./interfaces";
-export * from "./types";
+export * from "./auth";
+export * from "./ui";
